client: abstract generation
Dropping reliance on global variables allows creating clients on the
fly and working with multiple API specs in a single codebase.
